Restructure file-per-class. Remove path library
