+++ modules/lauf-store/src/core/watchable.ts
@@ -1,4 +1,4 @@
-import { Watchable, WatchableValue, Watcher } from "../types/watchable";
+import { Watchable, Watcher } from "../types/watchable";
 
 export class BasicWatchable<Value> implements Watchable<Value> {
   protected watchers: ReadonlyArray<Watcher<Value>>;
@@ -19,23 +19,5 @@ export class BasicWatchable<Value> implements Watchable<Value> {
         (candidate) => candidate !== watcher
       );
     };
-  };
-}
-
-export class BasicWatchableValue<Value>
-  extends BasicWatchable<Value>
-  implements WatchableValue<Value> {
-  protected value!: Value;
-  constructor(value: Value, watchers?: ReadonlyArray<Watcher<Value>>) {
-    super(watchers);
-    this.write(value);
-  }
-  write = (value: Value) => {
-    this.value = value;
-    this.notify(value);
-    return value;
-  };
-  read = () => {
-    return this.value;
   };
 }
